old-tests/submission/W3C/domhtml/HTMLFormElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - FORM</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-form-element.html#dom-form-elements" rel="help">
<link rel="author" title="NIST">
</head>
<body>

<form accept-charset="US-ASCII" action="./files/getData.pl" enctype="application/x-www-form-urlencoded" id="form1" method="post">
<p>
<textarea cols="20" name="text1" rows="7"></textarea>
<input name="submit1" type="submit" value="Submit">
<input name="submit2" type="reset" value="Reset">
</p>
</form>


<script>
var domtest = async_test("The elements attribute specifies a collection of all control element     in the form.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var elementnodeList;
      var testNode;
      var velements;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("form");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      elementnodeList = testNode.elements;

      velements = elementnodeList.length;

      assert_equals(velements,3);
       
    domtest.done();
}));

</script>
<p>The elements attribute specifies a collection of all control element
    in the form. 

    Retrieve the elements attribute and examine its value.</p>
<div id="log"></div>
</body></html>
